requirements.txt
flask
flask-cors
solana
pymongo
qrcode[pil]
python-dotenv
os
pillow
solders
